refactor(tui): descompresión Fase 1 — estabilidad TUI-first
- app.rs 2890 → app/{mod,state,types,cache} con Focus/Mode/NoteSort
  y NotePreviewCache como struct con is_valid_for (no más tupla 7)
- key_handlers.rs 8058 → handlers/theme.rs primer slice (theme picker)
  + scaffold handlers/ para splits futuros por focus
- tokio movido de workspace → shiki-desktop (no usado en tui/cli)
- scripts/gen-themes.py verifica triple-copy temas (Rust → CSS/JS)
  contra docs/css/styles.css y docs/js/main.js; OK 37 temas

Verificado: cargo check/clippy -D warnings/test 167 ok, fmt, doctor
