+++ shiki-cli/Cargo.toml
@@ -21,7 +21,6 @@ shiki-tui.workspace = true
 ratatui.workspace = true
 crossterm.workspace = true
 clap.workspace = true
-tokio.workspace = true
 anyhow.workspace = true
 libc.workspace = true
 tracing.workspace = true
