+++ shiki-tui/src/handlers/mod.rs
@@ -1,0 +1,12 @@
+pub mod theme;
+
+// Future splits will live here:
+// pub mod notebooks;
+// pub mod notes;
+// pub mod preview;
+// pub mod global;
+// pub mod editor;
+// pub mod settings;
+// pub mod git;
+// pub mod search;
+// pub mod drawer;
